requirements.txt
blinker==1.9.0
click==8.2.1
distro==1.9.0
Flask==3.1.1
gunicorn==23.0.0
itsdangerous==2.2.0
Jinja2==3.1.6
jpype1==1.5.2
MarkupSafe==3.0.2
numpy==2.2.6
packaging==25.0
pandas==2.2.3
python-dateutil==2.9.0.post0
pytz==2025.2
six==1.17.0
tabula-py==2.10.0
tzdata==2025.2
Werkzeug==3.1.3
